pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "django-template-reports"
version = "0.0.0"
description = "Generates reports (PPTX) from template files that are flexibly populated using the Django ORM, without hard-coding."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5,<6",
  "python-pptx>=1,<2",
  "swapper>=1,<2",
]

[project.urls]
Homepage = "https://github.com/gaussian/django-template-reports"
Issues = "https://github.com/gaussian/django-template-reports/issues"

[tool.black]
line-length = 90
extend-exclude = '''
/migrations/
'''
force-exclude = '''
/migrations/
'''

[tool.isort]
profile = "black"
